xwayland: Don't race/drop focus events
